- fixed a bug in windows
