+++ base/paniclog.win32.go
@@ -1,0 +1,7 @@
+// +build windows
+
+package jarvisbase
+
+func initPanicFile() error {
+	return nil
+}
